Fix the async keep-alive issue
Done by creating new async handles each time. The problem that was
async_close doesn't really close the handle until the next tick, however
before that point we might have already created a new handle,
overwriting the one which is about to be closed!

* Also revert back to having the bindings as the default export
* Add a function in debug mode to return active handles (i.e. those
which have a positive reference count)
* Fix some node 0.10 compile errors
*